move last fixture
